server: refactor storage add/dump
