+++ package.json
@@ -7,7 +7,7 @@
     "start": "node server",
     "nodemon": "node_modules/.bin/nodemon -w server/ server",
     "forever": "node_modules/.bin/nodemon -w server/ server",
-    "db-dump": "node server/storage-dump",
+    "db-dump": "fn=\"storage-dump-`date '+%Y-%m-%d--%H-%M-%S'`.json\"; echo saving into \"$fn\"; node server/storage-dump > \"$fn\"; echo saved into \"$fn\"",
     "db-add": "node server/storage-add"
   },
   "repository": {
